Move everything into src/ to prepare for tests
