+++ packages/faustwp-cli/src/healthCheck/index.ts
