+++ packages/faustwp-cli/src/index.ts
@@ -3,10 +3,10 @@
 import { spawnSync } from 'child_process';
 import dotenv from 'dotenv-flow';
 import { v4 as uuid } from 'uuid';
-import { debugLog, infoLog } from './lib/stdout/index.js';
-import { healthCheck } from './lib/healthCheck/index.js';
-import { generatePossibleTypes } from './lib/generatePossibleTypes.js';
-import { userConfig } from './lib/userConfig.js';
+import { debugLog, infoLog } from './stdout/index.js';
+import { healthCheck } from './healthCheck/index.js';
+import { generatePossibleTypes } from './generatePossibleTypes.js';
+import { userConfig } from './userConfig.js';
 import { getCliArgs, getWpSecret, getWpUrl, isDebug } from './utils/index.js';
 import {
   telemetryPrefsExist,
@@ -15,7 +15,7 @@ import {
   requestWPTelemetryData,
   shouldFireTelemetryEvent,
   sendTelemetryData,
-} from './lib/telemetry/index.js';
+} from './telemetry/index.js';
 
 // eslint-disable-next-line func-names, @typescript-eslint/no-floating-promises
 (async function () {
