+++ packages/faustwp-cli/src/stdout/index.ts
