+++ packages/faustwp-cli/src/stdout/styles.ts
