+++ packages/faustwp-cli/src/telemetry/doTelemetryPrefsExist.ts
