+++ packages/faustwp-cli/src/telemetry/index.ts
