+++ packages/faustwp-cli/src/telemetry/requestWPTelemetryData.ts
