+++ packages/faustwp-cli/src/telemetry/sendTelemetryData.ts
