+++ packages/faustwp-cli/src/telemetry/shouldFireTelemetryEvent.ts
@@ -1,5 +1,5 @@
 import { userConfig } from '../userConfig.js';
-import { getWpSecret } from '../../utils/index.js';
+import { getWpSecret } from '../utils/index.js';
 
 export function shouldFireTelemetryEvent(): boolean {
   const hasAnonymousId = userConfig.get('telemetry.anonymousId');
