+++ packages/faustwp-cli/src/utils/getCliArgs.ts
