+++ packages/faustwp-cli/src/utils/getGraphqlEndpoint.ts
