+++ packages/faustwp-cli/src/utils/getWpSecret.ts
