+++ packages/faustwp-cli/src/utils/getWpUrl.ts
