+++ packages/faustwp-cli/src/utils/index.ts
