+++ packages/faustwp-cli/src/utils/isDebug.ts
